feat: wip issue #3, can now login on database

+++ src/core/connectionInfo.cpp
@@ -1,6 +1,6 @@
-#include "connectionInfos.h"
+#include "connectionInfo.h"
 
-ConnectionInfo::ConnectionInfo(const QString& p_engine, const QString& p_host, const uint8_t p_port, const QString& p_database, const QString& p_username, const QString& p_password) :
+ConnectionInfo::ConnectionInfo(const QString& p_engine, const QString& p_host, const uint16_t p_port, const QString& p_database, const QString& p_username, const QString& p_password) :
     m_engine(p_engine), m_host(p_host), m_port(p_port), m_database(p_database), m_username(p_username), m_password(p_password)
 {
 
@@ -16,7 +16,7 @@ const QString& ConnectionInfo::getHost() const
     return m_host;
 }
 
-uint8_t ConnectionInfo::getPort() const
+uint16_t ConnectionInfo::getPort() const
 {
     return m_port;
 }
